requirements.txt
imageio
jupyterlab
matplotlib
notebook
numpy
scipy
opencv-contrib-python
pygame-ce
pandas
simpy